requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
cycler==0.10.0
decorator==4.4.0
falcon==2.0.0
gunicorn==19.9.0
hug==2.5.5
idna==2.8
imageio==2.5.0
importlib-metadata==0.18
kiwisolver==1.1.0
matplotlib==3.1.0
more-itertools==7.0.0
networkx==2.3
numpy==1.16.4
packaging==19.0
pandas==0.24.2
Pillow==6.0.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pytest==4.6.3
pytest-cov==2.7.1
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.1
